Move singleValueGenerator into its own folder
