+++ utils/games/averageInRange/averageInRangeGames.test.ts
@@ -4,7 +4,7 @@ import type { GlucoseRecord } from '~/types/glucoseRecord'
 import type { Thresholds } from '~/types/thresholds'
 import { createDate, getDayBefore, getMockGlucoseRecord } from '~/utils/test/testUtils'
 import { CurrentDayStatus } from '~/types/constants'
-import { generateSingleValueGlucoseRecords } from '~/utils/generators/singleValueGenerator'
+import { generateSingleValueGlucoseRecords } from '~/utils/generators/singleValue/singleValueGenerator'
 import type { DailyStreakStats } from '~/types/dailyStreakStats'
 
 const mockThresholds: Thresholds = {
